Version 5.0.1
Compiled with Java 21 & Vaadin 24.7.2.

SOChart is now extended from LitComponentWithSize.
The group Id for the artifact is renamed to `org.vaadin.addons.so`

Added proper event handling and various other features. GraphChart is rewritten and is not compatible with the earlier version. Also, there could be some other incompatibilities, and it is required to recompile your existing code. A new chart type is added: BoxplotChart.